Writing flow: absorb clipboard handler
