+++ packages/block-editor/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Deprecated `CopyHandler`, absorbed into `WritingFlow`.
+
 ## 12.10.0 (2023-09-20)
 
 - The Deprecated multiline prop on RichText will now fall back to using multiple
