+++ packages/block-editor/src/components/writing-flow/index.js
@@ -22,6 +22,7 @@ import useDragSelection from './use-drag-selection';
 import useSelectionObserver from './use-selection-observer';
 import useClickSelection from './use-click-selection';
 import useInput from './use-input';
+import useClipboardHandler from './use-clipboard-handler';
 import { store as blockEditorStore } from '../../store';
 
 export function useWritingFlow() {
@@ -35,6 +36,7 @@ export function useWritingFlow() {
 		before,
 		useMergeRefs( [
 			ref,
+			useClipboardHandler(),
 			useInput(),
 			useDragSelection(),
 			useSelectionObserver(),
